(PA-1833) Use full paths when loading agent-runtime-base
When vanagon uses `inherit_settings` to load settings from an external
project, if that external project uses `instance_eval` to load another
file, the path will be treated as relative to the original project
instead of the external project.

This commit modifies the `instance_eval`s in puppet-agent projects so
that they use full paths; This corrects the paths when other projects
use `inherit_settings` to copy settings from the runtime.