(PA-2055) Correct install_names for boost dylibs on macOS
Boost.Build's behavior around setting the install_name for dylibs on
macOS is not easily configurable. By default, it will hard-code the
relative build directory there, making the libraries unusable once
they're moved to the puppet libdir. Instead of dealing with this in a
jamfile somewhere, this commit adds a bash script to manually rewrite
the finshed dylibs' install_names so that they use @rpath instead of the
build directory, like other shared libs vendored in puppet-agent.